services/llm_runtime/requirements.txt
pydantic==2.9.2
pydantic-settings==2.5.2
grpcio==1.66.2
grpcio-tools==1.66.2
requests==2.32.3
langfuse==2.*